move s3_agent to a higher level package
